feat: add apache arrow lib

+++ 06-parquet-converter/internal/pkg/hotel/hotel.go
@@ -3,7 +3,7 @@ package hotel
 type Hotel struct {
 	Name   string  `parquet:"name"`
 	City   string  `parquet:"city"`
-	Review float64 `parquet:"review"`
+	Review float32 `parquet:"review"`
 }
 
 func NewHotel() *Hotel {
